+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### changed
+
+  - NuvlaBox status - new attribute nuvlabox-api-endpoint
+
 ## [4.2.11] - 2020-02-07
 
 ### changed