+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### changed
 
+  - Module component and application support private registries
   - NuvlaBox status - new attribute nuvlabox-api-endpoint
 
 ## [4.2.11] - 2020-02-07