+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### changed
 
+  - Credential - check registry credential
   - Module component and application support private registries
   - NuvlaBox - new actions check-api and reboot on resource
   - NuvlaBox status - new attribute nuvlabox-api-endpoint